Modifications to make the ModdedSongsManager handle reloads
This splits the construction of the manager into two parts. The first
part does the one-time load of the mods in TekaSongs, handling Fumen and
TJA songs. The second part, called "publishing" handles the process of
making the mods available to the music data of the game. This needs to
be called at startup and whenever a Reload() occurs, which is handled
with a postfix patch.

Manager functionality that was repeated in both the TJA and Fumen mods
was consolidated into a single method.

Tested this change with and without the SwapSongLanguages, verifying
that both Fumen and TJA mods were present, and that the
SwapSongLanguages mod worked as expected.