Fix song delay issue
Some songs have an initial BPM defined in the file, but the
first measure has a #BPMCHANGE command applied to it.

The delay for playing the start of the song needs to be computed from
the first measure BPM, not the overall BPM of the song.

This fix records the first measure BPM (is the same in all courses)
and uses that variable instead of song BPM to compute the delay.

Tested with
https://discord.com/channels/859872100860035113/1277751061343633559/1343050278894571553