Fix issues with saving mod data.
Rather than try to intercept the enumerable object inside the coroutine,
this hooks the Save UniTask to perform a pre-action step of hiding the
user data, and a post-action of restoring the hidden data.

The efficacy of this approach can be demonstrated by uncommenting the
Thread.Sleep() and verifying that the game just halts and will not do
anything else until the UniTask is complete, demonstrating there
are no race conditions with other parts of the game code now.